@@ -1,7 +1,7 @@
 {
   "name": "filesize",
   "description": "JavaScript library to generate a human readable String describing the file size",
-  "version": "10.0.6",
+  "version": "10.0.7",
   "homepage": "https://filesizejs.com",
   "author": "Jason Mulligan <jason.mulligan@avoidwork.com>",
   "repository": {
@@ -25,7 +25,7 @@
     "node": ">= 10.4.0"
   },
   "scripts": {
-    "build": "mkdir -p dist && rm -rf dist/* || rm dist/* || echo 'nothing in ./dist' && npm run rollup",
+    "build": "npm run rollup",
     "changelog": "auto-changelog -p",
     "coverage": "nyc npm run test",
     "lint": "eslint *.js src/*.js test/*.js",
@@ -36,12 +36,12 @@
     "types": "npx -p typescript tsc src/*.js --declaration --allowJs --emitDeclarationOnly --outDir types"
   },
   "devDependencies": {
+    "@rollup/plugin-terser": "^0.4.0",
     "auto-changelog": "^2.4.0",
-    "eslint": "^8.24.0",
-    "mocha": "^10.0.0",
-    "rollup": "^2.79.1",
-    "rollup-plugin-terser": "^7.0.2",
-    "typescript": "^4.8.4"
+    "eslint": "^8.37.0",
+    "mocha": "^10.2.0",
+    "rollup": "^3.20.2",
+    "typescript": "^5.0.2"
   },
   "keywords": [
     "file",
